api/rag_chatbot/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
uvicorn
fastapi